requirements.txt
# Core dependencies
python-dotenv==1.0.0
web3==6.15.1
eth-account==0.10.0
eth-typing==3.5.2
eth-utils==2.3.1

# Data processing and analysis
pandas==2.1.4
numpy==1.26.3
ccxt==4.1.87
python-binance==1.0.19
requests==2.31.0
aiohttp==3.9.3
websockets==12.0

# Machine learning and reinforcement learning
torch==2.1.2
transformers==4.37.2
stable-baselines3==2.2.1
gymnasium==0.29.1
scikit-learn==1.3.2

# API integrations
# coingecko-api==1.0.0
pycoingecko==3.1.0
coinmarketcap==5.0.3
# Custom implementations will be used instead of these packages
# web3-chainlink==0.1.1
# thegraph-client==0.1.0
gql==3.4.1

# DEX interactions
uniswap-python==0.7.1
# Custom implementations will be used instead of these packages
# curve-python==0.0.5
# paraswap-python==0.1.0
# kyberswap-python==0.1.0

# Flash loan providers
# Custom implementations will be used instead of these packages
# aave-v3-core==1.18.0
# dydx-v3-python==1.0.8

# Testing and monitoring
pytest==7.4.3
pytest-asyncio==0.21.1
prometheus-client==0.17.1
logging-plus==1.2.0

# Web interface
flask==2.3.3
flask-cors==4.0.0
dash==2.14.2
plotly==5.18.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
python-telegram-bot==20.7
